as-xinu: ポインタを受ける整数型をintptr_tに変更

+++ as/xinu/as13.c
@@ -11,7 +11,7 @@ short curfb[10] = {
 	};
 
 assem() {
-	int op, svop;
+	intptr_t op, svop;
 
   astart:
 	for(;;) {
