+++ as/xinu/as21.c
@@ -16,7 +16,7 @@ main(argc,argv) char **argv; {
 	int chr;
 
 	savint = signal(SIGINT, SIG_IGN);
-	if(((int)savint & 1) == 0) signal(SIGINT, saexit);
+	if(((intptr_t)savint & 1) == 0) signal(SIGINT, saexit);
 	argv++;
 	argc--;
 	while(argv[0][0] == '-') {
